Refactored application
